requirements.txt
fastapi==0.115.12
pydantic==2.11.3
python-dotenv==1.0.0
openai==1.75.0
uvicorn==0.34.2